feat: implement contact management APIs

+++ app/Services/ProfileService.php
@@ -1,10 +1,16 @@
 <?php
+
 namespace App\Services;
+
 use App\Http\Requests\ProfileRequest;
 use App\Repositories\ProfileRepository;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
+
 class ProfileService extends BaseService
 {
     protected StorageService $storageService;
+
     public function __construct(
         ProfileRepository $profileRepository,
         StorageService $storageService
@@ -12,12 +18,13 @@ class ProfileService extends BaseService
         parent::__construct($profileRepository);
         $this->storageService = $storageService;
     }
+
     public function getProfile()
     {
         $profile = $this->repository->getProfile();
         if ($profile) {
-            $profile->avatar = $this->storageService->url($profile->avatar);
-            $profile->cv_path = $this->storageService->url($profile->cv_path);
+            $profile->avatar = $profile->avatar ? $this->storageService->url($profile->avatar) : null;
+            $profile->cv_path = $profile->cv_path ? $this->storageService->url($profile->cv_path) : null;
         }
         return $profile;
     }
@@ -27,39 +34,48 @@ class ProfileService extends BaseService
         $profile = $this->repository->getProfile();
         $rawOldAvatar = $profile ? $profile->getRawOriginal('avatar') : null;
         $rawOldCv = $profile ? $profile->getRawOriginal('cv_path') : null;
-        if ($rawOldAvatar && str_contains($rawOldAvatar, 'http')) {
-            $rawOldAvatar = ltrim(parse_url($rawOldAvatar, PHP_URL_PATH), '/');
-            $rawOldAvatar = str_replace(config('filesystems.disks.s3.bucket') . '/', '', $rawOldAvatar);
-        }
-        if ($rawOldCv && str_contains($rawOldCv, 'http')) {
-            $rawOldCv = ltrim(parse_url($rawOldCv, PHP_URL_PATH), '/');
-            $rawOldCv = str_replace(config('filesystems.disks.s3.bucket') . '/', '', $rawOldCv);
-        }
 
         $data = $request->validated();
         unset($data['cv_file']);
+        $uploadedFiles = [];
 
-        if ($request->hasFile('avatar') && $request->file('avatar')->isValid()) {
-            if (!empty($rawOldAvatar)) {
-                $this->storageService->delete($rawOldAvatar);
+        try {
+            $updatedProfile = DB::transaction(function () use ($request, $profile, &$data, &$uploadedFiles) {
+
+                if ($request->hasFile('avatar') && $request->file('avatar')->isValid()) {
+                    $data['avatar'] = $this->storageService->upload($request->file('avatar'), 'avatars');
+                    $uploadedFiles[] = $data['avatar'];
+                } else {
+                    unset($data['avatar']);
+                }
+
+                if ($request->hasFile('cv_file') && $request->file('cv_file')->isValid()) {
+                    $data['cv_path'] = $this->storageService->upload($request->file('cv_file'), 'cvs');
+                    $uploadedFiles[] = $data['cv_path'];
+                }
+
+                return $this->repository->update($profile->id, $data);
+            });
+            if ($updatedProfile) {
+                if ($request->hasFile('avatar') && !empty($rawOldAvatar)) {
+                    $this->storageService->delete($rawOldAvatar);
+                }
+                if ($request->hasFile('cv_file') && !empty($rawOldCv)) {
+                    $this->storageService->delete($rawOldCv);
+                }
+
+                $updatedProfile->avatar = $updatedProfile->avatar ? $this->storageService->url($updatedProfile->avatar) : null;
+                $updatedProfile->cv_path = $updatedProfile->cv_path ? $this->storageService->url($updatedProfile->cv_path) : null;
             }
-            $data['avatar'] = $this->storageService->upload($request->file('avatar'), 'avatars');
-        } else {
-            unset($data['avatar']);
-        }
 
-        if ($request->hasFile('cv_file') && $request->file('cv_file')->isValid()) {
-            if (!empty($rawOldCv)) {
-                $this->storageService->delete($rawOldCv);
+            return $updatedProfile;
+        } catch (\Exception $e) {
+            foreach ($uploadedFiles as $filePath) {
+                $this->storageService->delete($filePath);
             }
-            $data['cv_path'] = $this->storageService->upload($request->file('cv_file'), 'cvs');
-        }
 
-        $updatedProfile = $this->repository->update($profile->id, $data);
-        if ($updatedProfile) {
-            $updatedProfile->avatar = $this->storageService->url($updatedProfile->avatar);
-            $updatedProfile->cv_path = $this->storageService->url($updatedProfile->cv_path);
+            Log::error('Profile Update Transaction Failed: ' . $e->getMessage());
+            throw $e;
         }
-        return $updatedProfile;
     }
 }
